First pass of moving all page options initialization to PO module
