+++ CancerGov/server/mock_views/social-share.html
@@ -648,14 +648,13 @@
                                                     <img alt="Resize font" src="/publishedcontent/images/images/spacer.gif" />
                                                 </a>
                                             </li>
-                                            <li class="po-print">
+                                            <li class="page-options--print">
                                                 <a title="Print" onclick="NCIAnalytics.PrintLink(this);" href="#">
                                                     <img alt="Print" src="/publishedcontent/images/images/spacer.gif" />
                                                 </a>
                                             </li>
-                                            <li class="po-email">
-                                                <a title="Email" href="/common/popUps/PopEmail.aspx?title=NCI+Events&amp;docurl=%2fnews-events%2fevents&amp;language=en&amp;a=181036Y0776&amp;b=f15387"
-                                                    onclick="NCIAnalytics.eMailLink(this); dynPopWindow(&#39;/common/popUps/PopEmail.aspx?title=NCI+Events&amp;docurl=%2fnews-events%2fevents&amp;language=en&amp;a=181036Y0776&amp;b=f15387&#39;, &#39;emailPopUp&#39;, &#39;height=525,width=492&#39;); return false;">
+                                            <li class="page-options--email">
+                                                <a title="Email" href="#">
                                                     <img alt="Email" src="/publishedcontent/images/images/spacer.gif" />
                                                 </a>
                                             </li>
